requirements.txt
Flask>=2.0
openai-whisper
torch
numpy
soundfile
scipy
transformers
ctransformers
protobuf
sentencepiece 